move sapling initializer functions to src/util
Doing this avoids code duplication
