+++ src/Makefile.bench.include
@@ -26,7 +26,9 @@ bench_bench_pivx_LDADD = \
   $(LIBLEVELDB) \
   $(LIBMEMENV) \
   $(LIBSECP256K1) \
-  $(LIBBITCOIN_ZEROCOIN)
+  $(LIBBITCOIN_ZEROCOIN) \
+  $(LIBRUSTZCASH) \
+  $(LIBZCASH_LIBS)
 
 if ENABLE_ZMQ
 bench_bench_pivx_LDADD += $(LIBBITCOIN_ZMQ) $(ZMQ_LIBS)
