+++ src/sapling/util.h
@@ -14,12 +14,6 @@
 #include <vector>
 #include <cstdint>
 
-// Init sapling library
-void initZKSNARKS();
-
-// Sapling network dir
-const fs::path &ZC_GetParamsDir();
-
 std::vector<unsigned char> convertIntToVectorLE(const uint64_t val_int);
 std::vector<bool> convertBytesVectorToVector(const std::vector<unsigned char>& bytes);
 uint64_t convertVectorToInt(const std::vector<bool>& v);
