+++ src/util.cpp
@@ -18,6 +18,8 @@
 #include "utilstrencodings.h"
 #include "utiltime.h"
 
+#include <librustzcash.h>
+
 #include <stdarg.h>
 #include <thread>
 
@@ -312,7 +314,90 @@ fs::path GetDefaultDataDir()
 
 static fs::path pathCached;
 static fs::path pathCachedNetSpecific;
+static fs::path zc_paramsPathCached;
 static RecursiveMutex csPathCached;
+
+static fs::path ZC_GetBaseParamsDir()
+{
+    // Copied from GetDefaultDataDir and adapter for zcash params.
+    // Windows < Vista: C:\Documents and Settings\Username\Application Data\PIVXParams
+    // Windows >= Vista: C:\Users\Username\AppData\Roaming\PIVXParams
+    // Mac: ~/Library/Application Support/PIVXParams
+    // Unix: ~/.pivx-params
+#ifdef WIN32
+    // Windows
+    return GetSpecialFolderPath(CSIDL_APPDATA) / "PIVXParams";
+#else
+    fs::path pathRet;
+    char* pszHome = getenv("HOME");
+    if (pszHome == NULL || strlen(pszHome) == 0)
+        pathRet = fs::path("/");
+    else
+        pathRet = fs::path(pszHome);
+#ifdef MAC_OSX
+    // Mac
+    pathRet /= "Library/Application Support";
+    TryCreateDirectory(pathRet);
+    return pathRet / "PIVXParams";
+#else
+    // Unix
+    return pathRet / ".pivx-params";
+#endif
+#endif
+}
+
+const fs::path &ZC_GetParamsDir()
+{
+    LOCK(csPathCached); // Reuse the same lock as upstream.
+
+    fs::path &path = zc_paramsPathCached;
+
+    // This can be called during exceptions by LogPrintf(), so we cache the
+    // value so we don't have to do memory allocations after that.
+    if (!path.empty())
+        return path;
+
+    path = ZC_GetBaseParamsDir();
+
+    return path;
+}
+
+void initZKSNARKS()
+{
+    fs::path sapling_spend = ZC_GetParamsDir() / "sapling-spend.params";
+    fs::path sapling_output = ZC_GetParamsDir() / "sapling-output.params";
+    fs::path sprout_groth16 = ZC_GetParamsDir() / "sprout-groth16.params";
+
+    if (!(fs::exists(sapling_spend) &&
+          fs::exists(sapling_output) &&
+          fs::exists(sprout_groth16)
+    )) {
+        throw std::runtime_error("Sapling params don't exist");
+    }
+
+    static_assert(
+        sizeof(fs::path::value_type) == sizeof(codeunit),
+        "librustzcash not configured correctly");
+    auto sapling_spend_str = sapling_spend.native();
+    auto sapling_output_str = sapling_output.native();
+    auto sprout_groth16_str = sprout_groth16.native();
+
+    //LogPrintf("Loading Sapling (Spend) parameters from %s\n", sapling_spend.string().c_str());
+
+    librustzcash_init_zksnark_params(
+        reinterpret_cast<const codeunit*>(sapling_spend_str.c_str()),
+        sapling_spend_str.length(),
+        "8270785a1a0d0bc77196f000ee6d221c9c9894f55307bd9357c3f0105d31ca63991ab91324160d8f53e2bbd3c2633a6eb8bdf5205d822e7f3f73edac51b2b70c",
+        reinterpret_cast<const codeunit*>(sapling_output_str.c_str()),
+        sapling_output_str.length(),
+        "657e3d38dbb5cb5e7dd2970e8b03d69b4787dd907285b5a7f0790dcc8072f60bf593b32cc2d1c030e00ff5ae64bf84c5c3beb84ddc841d48264b4a171744d028",
+        reinterpret_cast<const codeunit*>(sprout_groth16_str.c_str()),
+        sprout_groth16_str.length(),
+        "e9b238411bd6c0ec4791e9d04245ec350c9c5744f5610dfcce4365d5ca49dfefd5054e371842b3f88fa1b9d7e8e075249b3ebabd167fa8b0f3161292d36c180a"
+    );
+
+    //std::cout << "### Sapling params initialized ###" << std::endl;
+}
 
 const fs::path& GetDataDir(bool fNetSpecific)
 {
